Clean up code and move WifiAwareProxy to a separate file
