+++ AirCasting/SessionStopping/SyncTriggeringSesionStopperDecorator.swift
@@ -2,14 +2,10 @@
 //
 
 import Foundation
-import Resolver
 
 class SyncTriggeringSesionStopperDecorator: SessionStoppable {
     private let stoppable: SessionStoppable
     private let synchronizer: SessionSynchronizer
-    
-    @InjectedObject private var userSettings: UserSettings
-    @Injected private var networkChecker: NetworkChecker
     
     init(stoppable: SessionStoppable, synchronizer: SessionSynchronizer) {
         self.stoppable = stoppable
@@ -18,7 +14,6 @@ class SyncTriggeringSesionStopperDecorator: SessionStoppable {
     
     func stopSession() throws {
         try stoppable.stopSession()
-        
         synchronizer.triggerSynchronization()
     }
 }
